requirements.txt
django==4.1.3

django-cors-headers
# pymysql
# mysqlclient==2.1.0

pytest
pytest-django
coverage

uwsgi
bcrypt

zipfile36
django_q
Pillow
chardet
requests
python-magic;platform_system=="Linux"
python-magic-bin;platform_system=="Windows"

opencv-python
face_recognition
numpy
boost

openpyxl
psycopg2
